requirements.txt
python-dateutil
boto3
pytest
pyyaml
